+++ mayan/apps/mimetype/literals.py
@@ -1,0 +1,3 @@
+from __future__ import unicode_literals
+
+DEFAULT_MIMETYPE_FILE_READ_SIZE = 1024